requirements.txt
torch>=2.0.0
open_clip_torch>=2.20.0
Flask>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
